feat: implemented custom exceptions for errors

+++ src/main/java/com/api_banco/postgreSQL/service/ProdutoService.java
@@ -1,5 +1,6 @@
 package com.api_banco.postgreSQL.service;
 
+import com.api_banco.postgreSQL.exception.RecursoNaoEcontradoException;
 import com.api_banco.postgreSQL.model.Produto;
 import com.api_banco.postgreSQL.repository.ProdutoRepository;
 import org.springframework.stereotype.Service;
@@ -19,8 +20,9 @@ public class ProdutoService {
         return produtoRepository.findAll();
     }
 
-    public Optional<Produto> buscarPorId(Long id){
-        return produtoRepository.findById(id);
+    public Produto buscarPorId(Long id){
+        return produtoRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEcontradoException("produto com ID: " + id + " nao encontrado"));
     }
 
     public Produto salvarProduto(Produto produto){
@@ -28,6 +30,9 @@ public class ProdutoService {
     }
 
     public void deletarProduto(Long id){
+        if (!produtoRepository.existsById(id)){
+            throw new RecursoNaoEcontradoException("produto com ID: " + id + " nao encontrado");
+        }
         produtoRepository.deleteById(id);
     }
 }
